Added a test for equivalence
A new test confirming the result of the prior probability of freedom
calculation. This agrees with Jenny's Calculation in excel.

Also added a comparison study of the use of hyergeometric vs the
binomial approximation calculation for HSe. They largely agree below 10%
sampling fraction. There is, however, a region where dp and test Se are
high and sampling fraction is very low that the binomial results in a
HSe value up to 35% greater than the hypergeometric.