Fix bug in pert distribution when mu == x.mode
In the special case where the mean of the modified-PERT
distribution (mu) equals the mode specified in an argument then you
end up dividing by 0. A floating point math error causes the test for
equivalence to fail. This is solved by subtracting and testing that
the difference is very small.